+++ HISTORY.rst
@@ -3,6 +3,12 @@
 History
 -------
 
+v0.11.0 (2018-XX-XX)
+....................
+* make ``list``, ``tuple`` and ``set`` types stricter #86
+* **breaking change**: remove msgpack parsing #201
+* add ``FilePath`` and ``DirectoryPath`` types #10
+
 v0.10.0 (2018-06-11)
 ....................
 * add ``Config.allow_population_by_alias`` #160, thanks @bendemaree